3rdparty/mongo-c-driver-2.1.0/pyproject.toml
[project]
name = "mongo-c-driver"
version = "0.1.0"
description = "For development only."
requires-python = ">=3.10"
dependencies = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    { include-group = "docs" },
    { include-group = "evg" },
    { include-group = "format" },
]

format = [
    # tools/format.py
    "clang-format~=20.1.0",
]

docs = [
    # .evergreen/scripts/build-docs.sh
    "furo>=2023.5.20",
    "sphinx-design>=0.5.0",
    "sphinx>=7.1.1,<9.0",
]

evg = [
    # .evergreen/config_generator
    "packaging>=14.0",
    "pydantic>=2.8",
    "shrub-py>=3.7",

    # .evergreen/legacy_config_generator/evergreen_config_generator/__init__.py
    "yamlloader>=1.5",
]

[project.scripts]
mc-evg-generate = "config_generator.generate:main"

[tool.hatch.build.targets.wheel]
packages = [
    ".evergreen/config_generator",
]
